requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
neo4j==5.17.0
python-json-logger==2.0.7
pydantic==2.6.3 